Add TagThrottler class
